Refactor Admin state
